Move behaviours in another package;
Create classes for other behaviours.
