+++ src/main/java/it/unitn/ds1/Replica.java
@@ -16,6 +16,7 @@ import akka.actor.ActorRef;
 import akka.actor.Cancellable;
 import akka.actor.Props;
 
+import it.unitn.ds1.behaviours.ReplicaElectionBehaviour;
 import it.unitn.ds1.models.*;
 import it.unitn.ds1.models.administratives.JoinGroupMsg;
 import it.unitn.ds1.models.administratives.StartMsg;
