+++ src/main/java/it/unitn/ds1/behaviours/MessageTimeouts.java
@@ -1,0 +1,4 @@
+package it.unitn.ds1.behaviours;
+
+public class MessageTimeouts {
+}
